update convo list from message bus
